treewide: fix result type of BitmapBytePad() to size_t
sizes can't have negative numbers, so use the size_t here.
Also having to fix up lots of callers.

+++ Xext/xfixes/cursor.c
@@ -313,7 +313,7 @@ CopyCursorToImage(CursorPtr pCursor, CARD32 *image)
     {
         unsigned char *srcLine = pCursor->bits->source;
         unsigned char *mskLine = pCursor->bits->mask;
-        int stride = BitmapBytePad(width);
+        size_t stride = BitmapBytePad(width);
         int x, y;
         CARD32 fg, bg;
